+++ lib/scan.js
@@ -165,8 +165,8 @@ function initWriteQueue(options) {
   }
   done();
 },{
-  batchSize: options.historyCargoSize,
-  batchDelay: options.historyCargoDelay,
+  batchSize: 4,
+  batchDelay: 20000,
   batchDelayTimeout: historyBatchDelayTimeout, 
  });
