+++ package.json
@@ -42,7 +42,6 @@
     "chai": "^5.0.3",
     "esmock": "^2.6.3",
     "mocha": "^10.2.0",
-    "mock-fs": "^5.2.0",
     "nodemon": "^3.0.3",
     "sinon": "^17.0.1"
   }
